32. Update Menu Pelanggan

+++ ci4/app/Views/Kategori/select.php
@@ -40,7 +40,7 @@ if (isset($_GET['page_page'])) {
             <td><?= $no++ ?></td>
             <td><?= $value['kategori'] ?></td>
             <td><?= $value['keterangan']?></td>
-            <td><a href="<?= base_url()?>/admin/kategori/delete/<?= $value['idkategori']?>"><img src="<?= base_url('/icon/trash.svg')?>"></a>
+            <td><a onclick="return confirm('Hapus Data?')" href="<?= base_url()?>/admin/kategori/delete/<?= $value['idkategori']?>"><img src="<?= base_url('/icon/trash.svg')?>"></a>
             <a href="<?= base_url()?>/admin/kategori/find/<?= $value['idkategori']?>"><img src="<?= base_url('/icon/pencil.svg')?>"></a></td>
         </tr>
         <?php endforeach;?>
